requirements.txt
Flask==1.0.2
requests==2.11.1
pdftotext==2.1.0
mysqlclient==1.3.13
python-dotenv==0.9.1
